feat: 추상 클래스를 만들어 장,사 메서드 중복제거
